Added rigibody component and fix collision resolving

+++ src/core/controllers/Time.cpp
@@ -3,7 +3,7 @@
 
 namespace engine
 {
-    double Time::deltaTime = 0;
+    float Time::deltaTime = 0;
     std::chrono::time_point <std::chrono::_V2::system_clock> Time::previousTime = std::chrono::system_clock::now();
 
     Time::Time()
